+++ src/SIL.LCModel/Infrastructure/Impl/CmObjectIdentityMap.cs
@@ -55,7 +55,7 @@ namespace SIL.LCModel.Infrastructure.Impl
 		/// <summary>
 		/// Constructor
 		/// </summary>
-		internal IdentityMap(IFwMetaDataCacheManaged mdc)
+		public IdentityMap(IFwMetaDataCacheManaged mdc)
 		{
 			if (mdc == null) throw new ArgumentNullException("mdc");
 
@@ -82,8 +82,8 @@ namespace SIL.LCModel.Infrastructure.Impl
 
 		/// ------------------------------------------------------------------------------------
 		/// <summary>
-		/// Releases the memory stored in the IdentityMap. This is called automatically from
-		/// StructureMap when the StructureMap container is disposed.
+		/// Releases the memory stored in the IdentityMap. This is called automatically by the
+		/// DI container when the container is disposed.
 		/// </summary>
 		/// ------------------------------------------------------------------------------------
 		public void Dispose()