+++ src/SIL.LCModel/LcmGenerate/LcmServiceLocatorBootstrapper.vm.cs
@@ -7,10 +7,10 @@
 ## This file is used by the LcmGenerate task to generate the source code from the XMI
 ## database model.
 ## --------------------------------------------------------------------------------------------
+using Microsoft.Extensions.DependencyInjection;
 using SIL.LCModel.DomainImpl;
+using SIL.LCModel.Infrastructure;
 using SIL.LCModel.Infrastructure.Impl;
-using StructureMap;
-using StructureMap.Pipeline;
 
 namespace SIL.LCModel.IOC
 {
@@ -19,7 +19,11 @@ namespace SIL.LCModel.IOC
 	/// </summary>
 	internal partial class LcmServiceLocatorFactory
 	{
-		private static void AddFactories(Registry registry)
+		// Each type is registered by its concrete type (the actual singleton) with the
+		// interface registered as an alias resolving to that same singleton. This mirrors the
+		// previous StructureMap container, which allowed resolving a registered service by
+		// either its interface or its concrete implementation type.
+		private static void AddFactories(IServiceCollection services)
 		{
 #foreach($module in $lcmgenerate.Modules)
 #foreach($class in $module.Classes)
@@ -29,23 +33,20 @@ namespace SIL.LCModel.IOC
 #set( $classSfx = "Factory" )
 #end
 #if(!$class.IsAbstract)
-			registry
-				.For<I${class.Name}$classSfx>()
-				.LifecycleIs(new SingletonLifecycle())
-				.Use<${class.Name}$classSfx>();
+			services.AddSingleton<${class.Name}$classSfx>(sp => new ${class.Name}$classSfx(sp.GetRequiredService<LcmCache>()));
+			services.AddSingleton<I${class.Name}$classSfx>(sp => sp.GetRequiredService<${class.Name}$classSfx>());
 #end
 #end
 #end
 		}
 
-		private static void AddRepositories(Registry registry)
+		private static void AddRepositories(IServiceCollection services)
 		{
 #foreach($module in $lcmgenerate.Modules)
 #foreach($class in $module.Classes)
-			registry
-				.For<I${class.Name}Repository>()
-				.LifecycleIs(new SingletonLifecycle())
-				.Use<${class.Name}Repository>();
+			services.AddSingleton<${class.Name}Repository>(sp => new ${class.Name}Repository(
+				sp.GetRequiredService<LcmCache>(), sp.GetRequiredService<IDataReader>()));
+			services.AddSingleton<I${class.Name}Repository>(sp => sp.GetRequiredService<${class.Name}Repository>());
 #end
 #end
 		}